+++ src/app/components/habilitation/habilitation.component.html
@@ -38,7 +38,7 @@
         <pre>{{fonctions | json}}</pre>
     </div> -->
 
-    <mat-grid-list cols="2" rowHeight="2:1">
+    <mat-grid-list cols="3" rowHeight="3:2">
         <mat-grid-tile *ngFor="let f of fonctions; let j = index">
 
             <section class="example-section">
