+++ src/app/components/membre/membre.component.html
@@ -1,12 +1,12 @@
 <div style="margin-bottom: 10px; background-color: white;">
 
-    <mat-form-field class="example-form-field">
+    <!-- <mat-form-field class="example-form-field">
       <mat-label>Rechercher</mat-label>
       <input matInput type="text" (change)="searchMot(recherche)" [(ngModel)]="recherche">
       <button mat-button *ngIf="recherche" matSuffix mat-icon-button aria-label="Clear" (click)="initializeSearch()">
         <mat-icon>close</mat-icon>
       </button>
-    </mat-form-field>
+    </mat-form-field> -->
 
   <!-- <div style="text-align: right;">
     <button mat-raised-button color="primary">Liste des membres</button>
